[Silver I] Title: 곱셈, Time: 100 ms, Memory: 14084 KB -BaekjoonHub

+++ 백준/Silver/1629. 곱셈/곱셈.java
@@ -1,16 +1,16 @@
 class Main {
-    static long read() throws Exception {
-        long c, n = System.in.read() & 15;
+    static int read() throws Exception {
+        int c, n = System.in.read() & 15;
         while ((c = System.in.read()) > 47) n = (n << 3) + (n << 1) + (c & 15);
         return n;
     }
     
     public static void main(String[] args) throws Exception {
-        long a = read(), b = read(), c = read();
+        int a = read(), b = read(), c = read();
         System.out.println(rec(a, b, c));
     }
     
-    static long rec(long a, long b, long c) {
+    static long rec(int a, int b, int c) {
         if (b == 1) return a % c;
         long val = rec(a, b / 2, c);
         val = val * val % c;
